Refactory, documenting
